Rename file and fix import path for account classes in test_account.py
